MXS-2662: Add KeyManager configuration
The `key_manager` selects the key manager used by MaxScale. All keys are
managed under the same system, similar to how MariaDB has global key
management and encryption plugins.

The `key_manager_options` string is how the MasterKey implementations
will be configured. This is similar to how the cache storage modules are
configured. Not the most versatile system but it is adequate for the
small set of parameters that they (hopefully) will have.