MXS-2662: Read and write START_ENCRYPTION_EVENTs
All events after a START_ENCRYPTION_EVENT will be encrypted. When one is
detected, it is a signal to go into a mode where all events are assumed
to be encrypted. The event must be read also by the writing thread as it
may continue processing a binlog after a restart.