MXS-2662: Read and write encrypted events
With this commit, basic replication of encrypted events works. The only
exception is the GTID scan that is done to find the file position where
replication start. This means that real replication clients do not work
but `mysqlbinlog` does as it uses file-and-position and effectively asks
for the whole history.