MXS-4054 Use string_view in qc_get_[database|table]_names()
In Gcc, creating a string_view from a nullptr works and you get
an empty string_view, but the standard does not currently
guarantee that. Apparently there is a proposal for changing the
standard so that both a string and a string_view can be created
from a nullptr. But until that, it has to be checked if that is
a possibility.