MXS-2662: Move file read/write code into maxbase
Both the REST API code as well as the KeyManager both need to read files
from disk. Given that this is a relatively common pattern, moving them
into maxbase makes sense.

The mxb::load_file function takes the container type as one of its
template arguments to allow it to be used with both binary files
(usually `std::vector<uint8_t>`) and text files (almost always
`std::string`).

The `mxb::save_file` function is a modified version of the code found in
config_runtime.cc.  The writing of the file does not need to be a
template as a pointer to the data and its size can be easily derived
from the data. An overload for strings is needed as writing strings into
files seems to be a very common pattern in MaxScale.