MXS-2662: Add MasterKeyBase
The common task of fetching keys and using them to encrypt the keystore
file is shared across all planned MasterKey implementations.

The exception to this might be the AWS KMS that offers encryption and
decryption APIs which means we could use to perform the operations
without ever seeing the customer master key.