MXS-4054 Store the canonical version in the parsing info
Now the canonical version of the SQL in the GWBUF is stored in the
parsing info structure. All fields etc. that now are separately
allocated char*s will be converted into std::string_views pointing
into that canonical string.

Further, the canonical string will be exported as an std::string_view,
meaning that the query cache eventually can use that as its key. Thus,
there will be no duplication of data.