MXS-4090: Add failsafe for unexpected responses
If the backend protocol itself is broken, as is seen in MXS-4090, an
unexpected response can still end up being routed to the routers. To
handle the cases where an unexpected result will cause a crash, we need
code in readwritesplit that kills the session if it was't expecting a
result. This should prevent the crash while also allowing better debug
information to be generated in the hopes that it will eventually be
fixed.