MXS-4054 Store prepare name as string_view
This does not necessarily save any memory. If the statement is
a prepared statement, then there will be no extra allocation for
the name itself and memory will be saved. However, as most
statements are not prepare statements and as sizeof(char*) is 8
and sizeof(string_view) is 16, the cost of using string_view
instead of char* will in general be larger. But, in the big
scheme of things it's nicer to consistently use string_view.