MXS-4054 Count size correctly
Also order member variables slightly.

Note for the future. The data structure used while parsing and
the data structure used for storing the classification information
should be distinct. And there could be different kind of structures
for different kind of statements. And as the needed size could be
figured out at the point the classification data needs to be stored
the needed memory could be allocated in one single chunk so that
there would be no fragmentation.