MXS-2662: Make KeyManager runtime configurable
The key_manager and key_manager_options can now be modified at
runtime. This allows the "client master key" to be changed to a
different one which causes the keystore file to be encrypted with a
different key. This effectively rotates the keystore encryption key but
the actual files are still encrypted using the same keys.

As the KeyManager instance now changes at runtime, the retrieval of the
current key manager must be done under a lock and the instance must be
stored in a shared_ptr.