MXS-2662: Add HashiCorp Vault key manager
The key manager reads data from a Vault server. The code should be
compatible with the MariaDB version of the plugin which means it should
be possible to use the same encryption keys with the MariaDB server and
MaxScale.

Currently the key manager has no timeouts and the key retrieval is
attempted only once.

The libvault library requires full C++17 support, including the
filesystem library, which means it cannot be build on Ubuntu 18.04 as it
doesn't have GCC 8. For CentOS 7, the devtoolset must be upgraded from 7
to 9 to enable GCC 9.