MXS-4152: Split maps into two levels
Partitioning the shard map first by the database and then the table
allows all tables to be found with two lookups. This greatly speeds up
the schemarouter that previously iterated over all the results to find
the targets that have the tables in question.

A sysbench run with 50k tables and a debug build previously resulted in
very bad performance but with the new shard map it is roughly equal to
the performance of readwritesplit.