MXS-4105 Handle timeout errors properly
If the VM where Redis runs is powered off or if the docker
container where Redis is run is paused, then Redis commands will
just hang unless a timeout has been specified.

In that case, reconnecting will appear to succeed, but the next
command will then hang for the specified timeout, which will then
stall the client. So as not to make every single command thereafter
take timeout seconds, a reconnection will be attempted ever more
rarely and finally once per minute until it succeeds.