Fix result processing for database maps
The second field of each row is expected to be empty in some cases. The
result contains the combination of both tables and databases, the latter
of which has an empty string as its field. This was an oversight that
was added in the merge from 6.

This fixes most of the schemarouter tests that were failing.