MXS-4142 Prune aliases when configuring using json.
The runtime reconfiguration is done so that first the existing
configuration is obtained as a json object from the config, then
that is merged with the json object coming from the user, and
finally the config is configured with that json object.

The json object obtained from the config will not contain aliases,
but the one from the user may. To ensure that the user provided
value overrides, aliases used by the user are first removed and
finally added using the actual and not the alias name.