MXS-4025 Handle clear all sessions when deleting a worksheet
All sessions of a worksheet need to be cleared. i.e. its memory
states need to be reset and its binding connection need to be
disconnected.

This also stops the bubbling of the clicking event on the delete
worksheet icon button, as a result, clicking on that button,
it won't change the current active worksheet. This helps to
clear all sessions easier.