MXS-4145 Implement nosql::UserManagerMariaDB
With this commit, the MariaDB based user manager is functional.
For authentication and any user related action (e.g. list users),
the server will be accessed in a serialized manner. This should
not be a significant penalty provided users are not manipulated
often (likely) and NoSQL sessions are relatively long-lived (also
likely), especially since this is still a technical preview.
The JSON data containing sha1(password) is not yet encrypted.