MXS-2662: Fix seeking into unencrypted binlogs
If the binlog is not encrypted, the first two events of the file would
always be discarded. If the binlog was unencrypted, the second event
would be the GTID_LIST event which caused the GTID search code to never
see it.

Added an info message for the case where the GTIDs are not found. This
should help track cases like this where a requested GTID was not found.