Fix value stringification
Any values that loosely evaluated to false were converted to empty
strings. This caused the number zero to also be converted into an empty
string. As both undefined and null compare equal to null, the value can
be compared to null to test if we should use it or replace it with an
empty string.