Fix config_sync
For whatever reason, the test modified the second MaxScale but expected
it to take place on the first.

The user data cannot also be compared naively: the data contains
timestamps which can be different depending on the exact time the logins
take place. For this reason, only the account type is compared.