MXS-4171: Check for modifiable parameters
If a parameter changes value after the initial configuration and it was
not declared with Param::AT_RUNTIME, it must not be allowed to
change. To prevent partial configuration changes, the checks must be
done first.

The check for unknown parameters should also be done before the
parameters are modified but the validation done by the Specification
class already does is.