MXS-2662: Allow whitespace to be in Vault secrets
An easy and secure way of generating a new key is to pipe the output of
`openssl rand` into the vault command line client. OpenSSL adds a
newline into the key and that newline is included in the final data
stored in the Vault server. To accommodate this use-case, whitespace in
the keys should simply be ignored.