MXS-2662: Remove internal keyring
With the addition of nosqlprotocol remote authentication, consistent
encryption is needed. Simplifying the KeyManager interface to just fetch
keys by their IDs and versions allows this to be done with next to no
modifications to any modules that use the current local keyring. If a
local keyring is needed, perhaps as an extension to the
key_manager_file, it can be done internally by the MasterKey
implementation.

The task of managing versioned keys can be offloaded to the vault key
manager in the cases where such is needed. Given that most cloud key
management systems provide some form of versioning, this can be
relatively easily extended.

The one downside of this simplification is that the performance and
versatility of the key manager largely depends on the
implementation. This makes it harder to provide consistent behavior but
this is a perfectly acceptable approach when simplicity and versatility
are more important.