MXS-4145: Use KeyManager in nosqlprotocol
The encryption key is now retrieved from the key manager when the
protocol object is created. If the key manager is configured for
MaxScale and shared authentication is enabled, the passwords will be
automatically encrypted using it. One usability improvement that could
be added is to add a boolean that controls this independently of the
global encryption configuration.

Currently all new users are created using the latest encryption key
available in the key manager. This does come at the cost of being slower
due to a need to query the key manager which can involve network traffic
with remote key manager implementations. Luckily this can be optimized
on the key manager side by caching the latest result.

Decryption of older versions is done using the current local encryption
key if the versions are the same. This means that if the user was
created with the latest encryption key, there's no extra performance
cost to be paid. If the user was created with an older version, the key
manager is queried for that specific key version. This could also be a
place where the data could be upgraded to use a newer key version.