Fix session idle time calculation
The idle time used only the last read when both reads and writes must be
inspected.

The code in 2.5 lacks the ability to detect idleness of a session and
thus ongoing long queries (e.g. `ALTER TABLE`) are treated as being
idle. However, once the queries complete the timing information will
again be correct.