MXS-4178 Replace the usage of `localStorage` with `localForage`
localStorage has a size limit of 5MB for most modern browsers
unless the user increases it.
There is a better one which is IndexedDB. It's suitable for storing
larger amounts of structured data. It can store both objects and
key-value pairs, so it saves time converting objects to strings
that are needed when using localStorage. Besides, its maximum size
limit is also larger than localStorage.

This installs `localForage` which will automatically use
localStorage in browsers if no IndexedDB or WebSQL support.
This library internally uses `IndexedDB` with a simple, localStorage-like API.