MXS-4171: Extend MaxCtrl test case
Extended the test to cover modifications to static parameters. The test
also reveals the problem where aliased module names do not compare
equal.

Modules with aliases to new names (e.g. mysqlmon → mariadbmon) not being
treated as equal causes problems with the configuration
synchronization. During the synchronization, the whole configuration of
the object is patched over the old one and the addition of the checks
for runtime modifications to static parameters causes the sync to fail
if an old alias name was used.