Skip MaxScale startup in config_test
The test doesn't have the REPL_BACKEND label which means the
configuration processing code for tests ignores the `###server_line###`
section. This causes the configuration to be invalid which causes the
startup to fail. The test itself passes as it ignores this and returns a
different value but the logged messages make it look like the test
failed. Skipping the startup both fixes the output and makes the test
slightly faster.