Fix mxs922_server
Doing a DNS lookup on an unknown hostname can take a long time. To speed
things up and to make the test work correctly, use a known good IP
address. The test itself doesn't need an address that works, as long as
there's no server on that port even localhost works.