MXS-4171: Compare equality using native types
Doing the comparisons after the conversion from strings and JSON into
native types makes it possible for any aliased values to be resolved to
new ones. This fixes the problem where a no-op modification was rejected
due to two JSON values not comparing equal when the native values would
compare equal.