MXS-3723 Simplifying the process of opening file
This removes the option to open file in a new query tab as
it causes hassle when storing the file handle and it doesn't actually
bring much value to the user.

The `Open Script` confirm dialog will have three buttons:
`Cancel`, `Don't Save` and `Save`. The dialog will be shown only
when the current session query tab has unsaved changes to the opened
(uploaded) file.
* Clicking on `Don't Save` button will ignore the action to save the file to
the user's local device (or download the file if the browsers don't support
File System Access API) and open the chosen file.
* Clicking on `Save` button will save (or download) the file and open the
chosen file.

This commit changes the labels for those buttons and add `file_handle` state
to the `editor` state module.  This state will be later used for detecting
unsaved changes.