Remove unwanted data from admin users JSON
When used in configuration synchronization, the users must be comparable
as JSON to detect if a no-op change is being made. As the timestamps are
there only as extra information, they should be remove when the
resulting JSON is for internal use.

This fixes a problem that caused the config_sync test to fail.