Unmask passwords higher up in the call chain
The JSON output of a mxs::config::Configuration is used in some cases to
create a complete configuration from a partial set of parameter. This
has to be done as there is no distinction between the initial
configuration and the update to one with the current configuration
system. As a result of this, the password masking should be done as soon
as possible but only in functions that expect this to happen.