MXS-4025 Remove the `WORKSHEET` connection binding type
There is no need to keep this type. In fact, by keeping this type,
the user can not delete the first session tab and the user may get
confused. This type was kept in order to keep the "default" connection
data even when the worksheet is deleted, however, this doesn't bring
much benefits to the user as the connections are bound to the worksheet,
once the worksheet is deleted, its connections should be closed and removed.

This removes that connection binding type.