MXS-3723 Allow deleting the last session of the worksheet
The user may want to delete the last session to have a fresh session
tab. e.g. The user opens files on several session tabs and then wants
to close them all. The session tabs are deleted normally except for
the last session tab because it holds the connection-related data.
So it needs special treatment. When deleting the last session of
the worksheet, it actually clears all data of that session but keeps
the connection-related data. i.e. `active_sql_conn`.

This adds `handleClearTheLastSession` action to handle
that case.