MXS-4025 Fix undefined connection id for the `disconnect` action
The `disconnectAll` action iterates through each connection in
`sql_conns` and the `disconnect` action. The connection id may
be undefined if it has already been deleted by the previous iteration.
Because the `disconnect` action also delete the cloned connection of
the provided connection.

This checks for the existence of the connection id before deleting it.