MXS-4025 Fix duplicated navigation and wrong url when deleting a worksheet
When entering the query editor page, it shows duplicated navigation error
in the console because the `updateRoute` is called twice via
the `active_wke_id` watcher and `$route` watcher.

When deleting a worksheet, the url should be updated to
the previous tab but it updates to a wrong url. This adds
`prevent` modifier to stop this behavior.