MXS-4181: Replace tmp_rsa callback with ecdh_auto
The ephemeral RSA key exchange isn't used as it is supposedly a weak
method for key exchange. Additionally, the ephemeral RSA key exchange
was removed in OpenSSL 1.1.1 as was the requirement to manually turn on
automatic ECDH curve selection.

CentOS 7 with FIPS mode enabled would not use the ECDH ciphers at all
due to the missing call to SSL_CTX_set_ecdh_auto. The OpenSSL 1.0.2
documentation states:

  Newer applications should just call `SSL_CTX_set_ecdh_auto` and they
  will automatically support ECDH using the most appropriate shared
  curve.