MXS-4185 Update bootstrap servers with domain name properly
The dynamic servers will all have their addresses specified as
IP-addresses while bootstrap servers may have them defined using
domain names. To ensure that the bootstrap node is found, a
name lookup is made. If the address is already an IP-address,
that will be returned, so it works even if the bootstrap node has
been defined using an IP-address.