MXS-3723 Conditionally showing the `Save Script` icon button
The button should only be shown for browsers that support the
File System Access API and its click handler should handle one thing.
On the other hand, the `Save Script As` button should be shown for
any browsers. If it supports the API, it uses `saveFileAs` method
to open a dialog to choose the location to save as a new file,
otherwise, it downloads the file.

This also adjusts the name of some methods in the component to
be consistent with other methods.