MXS-2662: Add key_manager_vault test case
The test is largely the same as the one for key_manager_file. The only
real differences are the setup and the configuration they use.

Also added an overloads with std::string as the arguments to
copy_to_node and copy_from_node. This makes it easier to use them with
constructed paths.