MXS-4172: Improve correct_packet_sequence()
The code now uses the mxs::Buffer iterators to modify the sequence
numbers. This is more efficient than using gwbuf_copy_data and
gwbuf_byte_pointer as these always require iteration from the beginning
to modify the next packet.

The use of mxs::Buffer iterators also has another benefit: if the buffer
contains malformed data or some other part of the code ends up passing a
partial packet to the router, the iteration with advance() won't go past
the end iterator. This means that an attempt to read past the end of the
buffer is intercepted correctly and the iteration exists immediately.