MXS-2662: Fix reading of unencrypted binlogs
If the latest binlog stored on disk is not encrypted and encryption is
enabled after a restart, an attempt to write an encrypted STOP_EVENT
would be made. This happens because the code checked for a non-empty key
ID when it should check for a non-empty encryption context. Unencryped
events can be written into binlogs even if encryption is enabled.

Added the missing catch for encryption errors into the Writer and
changed the encryption related errors from BinlogWriteError into
EncryptionError. This will correctly stop the writing of the files if a
key cannot be retrieved.