MXS-3216 Make the `page-header` of the ClusterDetail page reusable
Monitor operations such as reset-replication, release-locks and failover
should be also available in the MonitorDetail page. So that later when
columnstore operations are added, it would be available in both
pages.

This made the `page-header` component reusable. The next commit would
take that component into use for the MonitorDetail page.