Make kill_query more stable
The query it used sometimes ended up not being interrupted due to the
fact that it was a non-atomic block and the KILL QUERY could end up
being executed at a point where it would not do anything.

Using a massive sequence table and max_statement_time appears to provide
a more stable alternative for a non-atomic block.