MXS-3216 Reduce the complexity of argument passing  in the `checkAsyncCmdRes` action
There's a need to add a parameter called `showSnackbar` to
conditionally show the result in the snackbar when the `async-cs-get-status`
command is added in the GUI. The return meta is an object containing
columnstore node status and it will be presented in a table not a snackbar.

The `checkAsyncCmdRes` action conditionally calls two handlers and one of them
even has a recursive call to the `checkAsyncCmdRes` action. This causes
complexity when adding a new parameter as it has to be passed down
to those handlers.

This keeps the parameters as an object instead of destructuring
the parameters so that it can be passed as an object argument.