Stop HttpSql cleanup thread on shutdown
If the shutdown is done normally, the MainWorker will end up calling
`HttpSql::stop_cleanup()` which stops the cleanup thread. However, if
the maxscale process receives two SIGTERM signals before it completes
the shutdown, it calls `exit()` which static objects to be destroyed.

The `ConnectionManager` class expects the thread to be stopped when its
destructor is called. This will not be true if the shutdown happens due
to the `exit()` call which in turn would trigger a debug assertion. For
release mode builds, this would throw an exception for a thread that is
destroyed but not joined. Both cases resulted in unsuccessful shutdown
even if the expectation was for the process to exit normally.

Always calling `stop_cleanup_thread()` in the destructor makes sure this
case is also handled.