Fix test_kafkacdc
The test didn't check which events were being sent, just that any events
were sent. The log messages show that binlog events unrelated to this
test were being sent to Kafka which caused the first part of the test to
pass but caused the second part to fail as it didn't have the right
content. Explicitly setting the starting GTID position in the file is a
more stable method of guaranteeing that only those events that the test
inserts are being processed.

Converted the `tprintf` calls to `log_printf` calls to show which events
relate to which parts of the test. This should make it easier to figure
out which part of the test output relates to which line in the MaxScale
log.