MXS-3216 Show snackbar while stopping/starting cluster
It usually takes time to finish the operation, so it's better to have
a snackbar while looping to get the command result.

This enables `showSnackbar` back for stop/start cluster operations.
It was disabled in order to hide the snackbar using meta response in
the `handleAsyncCmdDone` as  the meta response is a timestamp value
for ColumnStore operations which is not user-friendly to show.

This adds `custAsyncCmdDone` to show a custom snackbar. By doing this,
the pending/running snackbar can be shown to inform the user that
the operation is still running.