MXS-3216 Add `shouldFetchCsStatus` props to `monitor-page-header`
The MonitorDetail page dispatches the `handleFetchCsStatus` action
when the page it's created to show its data in the table, so there
is no need to call it in the `monitor-page-header`. However,
when the `monitor-page-header` is used in the ClusterDetail page,
it doesn't dispatch that action but it still needs the ColumnStore
status data to conditionally disable start/stop cluster operation and
detect if the operation is successfully done.

This adds the mentioned props to handle that.