Fix the `on-count-done` event in the `refresh-rate` component
If the handler for the event takes time to finish, the workerTimer
onmessage callback won't actually wait to finish it because it
internally uses `setInterval`. This causes conflicts and leads
to several `on-count-done` event emits at the same time.

This adds a `paused` state to conditionally call the `updateCountTime`
method. So the workerTimer still runs every 1s but the countdown
value won't be updated.