Make sanity_check more stable
The test case for MXS-3915 sends a set of session commands and then
immediately sends a SELECT which is assumed to not be routed to the
master. SELECT statements can be sent to the master server if none of
the slave server have responded to the session commands. The test
doesn't take this into account and it should repeat the read multiple
times to give the backends time to complete the commands. This still
covers the original test case where the transaction state was left open
and all future SELECT statements were treated as being inside of a
transaction.