MXS-2542 Add more checks to rebuild
Check that "sudo rm" and "sudo chown" are ran on the expected
directory. Run cleanup on success.

When preparing to start replication, update server status as if
a monitor tick is going on. This ensures that server version
info is updated even for servers that were down when monitor
started.