MXS-4203: Make use_sql_variables_in=all smarter
The current resultset processing and session command code can handle
commands that return resultsets as responses. This means the default
behavior of returning errors is no longer needed. This change could've
been done in 2.5 but since it changes behavior, it's better to do it in
a new version.

The variable should still be kept around to make it possible to prevent
resultsets from being counted as session commands.