MXS-3216 Add `Remove node from cluster` and `Add node to cluster` cs operations
This adds `Hostname/IP` input for selecting a node or entering the IP of
the node to be removed from and added into the ColumnStore cluster.

When adding node to the cluster, the suggested nodes are the MaxScale
server names. On the other hand, when removing the node from the CS cluster,
the suggested nodes are the key of each node in the meta response of
the `async-cs-get-status` cmd. Apart from the suggested nodes, the user
can enter the hostname or IP of the node manually.