Fix nested parameter prefixes
The prefix given to the module was expected to be the same as the value
given to the protocol parameter. This means that with
`protocol=MariaDBClient` the prefix is `mariadbclient` whereas with the
real module name, MariaDBProtocol, the prefix is `mariadbprotocol`. To
make matters worse, there's also aliases for `mysqlclient` and `mariadb`
that all resolve to the same module name. To make it easier to keep it
consistent across all documentation, the real module name should be used
as the prefix, not the pre-aliased value.

This also fixes the case where the prefix was not forced to lowercase
when it was read from JSON.

This doesn't have any functional differences when it comes to the
MariaDBProtocol as it has no parameters but this fix is required if
parameters are ever added to it.