MXS-4025 Fix leftover cloned connections when deleting a connection
The cloned connection is created by creating a new session tab,
it clones the connection bound to the last session. When the `disconnect`
action is called in the `connection-manager` component, the connection being
deleted is the connection bound to the first session of the worksheet.
As a result, it failed to finds all cloned connections by using the
`clone_of_conn_id` attribute of the cloned connection.

This changes `getWkeLastSessConnByWkeId` to `getWkeFirstSessConnByWkeId`
so that when cloning the connection, it will always uses the connection
of the first session tab.