MXS-4025 Fix `validatingConn` for failing to update active_sql_conn correctly
There is an edge case where this action fails to update `active_sql_conn`.
It happens when there are already some worksheets, the user leaves
the page and then navigate back to the query editor by using an
absolute link or clicking the button to go to the query editor of a
server in the ClusterDetail page. e.g. `/query/servers/server_1`.
The connection for this server is not yet created, so a new blank
worksheet is created. The `state.active_sql_conn` is a memory state
so when the new worksheet is added, this state is assigned to it.

This replaces the use of `state.active_sql_conn` in the `validatinConn`
action  with the `active_sql_conn` attribute of the active session.