MXS-4206 Extend `split-pane` to revert rendering panes
The priority of panes is a matter in Vue. In the `split-pane`
component, Vue creates the `pane-left` first, then the resizer and
the `pane-right`. It's similar to event propagation. There is a
case where the right pane needs to be handled first which will be
introduced in the upcoming commit.

This adds `revertRender` props to the `split-pane` component to
conditionally revert the render priority of the panes.