MXS-4206 Use `revertRender` in the `worksheet` component
This reverts the render priority of panes of the worksheet.
The `session-tabs` is now created and rendered first so that
the `active_sql_conn` can be available before the `sidebar-container`
is created and render.

This is done to make it possible to move the `schemaSidebar/initialFetch`
action calls from the `conn-man-ctr` to the `sidebar-container`. The
`conn-man-ctr` component should only handle connection related actions.

This also removes unused map states and getters.