MXS-4209: Close LocalClients on first response
When the response to the KILL command is received, the client can be
immediately closed. By using the cherry-picked LocalClient notification
callbacks from 6.4, the code can now properly close the connections even
when the TCP connection won't be closed. This also has the added benefit
of making the KILL commands faster as they are now fully event driven
and don't poll for the completion.