MXS-4206 Break `openConnect` action into smaller actions
When opening a connection, it handles complex situations when there
are multiple session tabs, it needs to bind the created connection
to the first session tab then clone that connection and bind it
to other session tabs. So, the action should be divided into smaller
actions to improve readability and reusability.

These smaller actions will be later reused to handle selecing
existing connections in the `conn-man-ctr` dropdown.