MXS-4206 Handle delete the session having the default connection
When deleting the session having the default connection, the connection
bound to it shouldn't be disconnected, because that makes it impossible to
disconnect its cloned connections when the user uses the `conn-man-ctr`
dropdown. The `conn-man-ctr` dropdown shows only the default connections,
so when a connection is disconnected in the dropdown, the `disconnect`
action finds it cloned by using the `clone_of_conn_id` attribtue. This
wasn't handled properly when the multi-session tabs was introduced in
MXS-4025.

This prevents the call to disconnect the default connection when deleting
a session.