MXS-4025 Handle set active db when adding a new session
The sidebar schemas tree data is the same regardless of the active
session tab. This is because, all session tabs connects to the same
server host, so its schemas are the same. However, when creating a
session connection, its active db is null.

To simplify, the new session connection will automatically set
the active db as the previous session. In other words, the active db
value will be the same for all session tabs. In fact, this is how
MySql Workbench does it.