MXS-4206 Change `charset_collation_map, def_db_charset_map` to object
With the recent changes related to layering containers and
presentational components, alter table feature is broken due to the
reactivity of those states. This is because Vue 2 doesn't support
Map object. It works before because the `ddl-editor-container`
is rendered after the sidebar is renderer so it's always 1 tick
slower. But with the recent changes, it encounters error when reading
those states.

This changes the type of those states to object.