MXS-4206 Simplify the process of open DDL editor to alter table
When the editor mode is `DDL_EDITOR`, there is a conflict when deleting
the session with unsaved changes, it fails to open the confirmation
dialog as the dialog is placed in the `txt-editor-container` not
`ddl-editor-container`.  To solve this, when altering table, it
will create a new session to display the `ddl-editor-container`.

This commit moves the mutations for preparing states for the ddl
editor from the `schema-tree-ctr` to `sidebar-ctr` to easily
handle the creation of the new session and bind states to the new
created session.