MXS-4206 Keep alter table data even after pressing f5 - part 1
Now the DDL editor is opened in a new session tab, when
refreshing the browser, the editor is gone leaving the
session tab with a confusing tab name. i.e. `ALTER table_name`.
This is because the editor mode and the table creation info are
stored in memory.

This moves the `curr_editor_mode_map` state from memState to syncState
so that it can be persisted to each session object. This also adds
`getIsTxtEditor` and `getIsDDLEditor` getters to the `editor` state
module as there are several components need these values.