MXS-4217: Make JWT algorithm configurable
This adds the ability to configure the algorithm that is used to sign
the JWT that the REST-API /v1/auth endpoint creates. The asymmetric
algorithms use the TLS certificates used by the REST-API: this means
that the certificates must be of the correct type (e.g. RSA for RS and
PS, ECDSA for ES) for them to work.