MXS-4217: Use symmetric keys from key manager
If a symmetric signature algorithm is used, the key used with it can now
be retrieved from the key manager. This allows multiple MaxScale
instances to share the same set of signature keys.

Currently the key is only read on startup but the `maxctrl reload tls`
command could be made to reload the key if needed.