MXS-4217: Use UUIDs for SQL connections
This helps avoid the case where two different MaxScales end up using the
same ID for the connection which will be a problem now that the JWTs are
valid across MaxScale instances.

The functionality of the connection IDs remains the same with one
exception: the connection ID is now separated by a period instead of a
hyphen as the UUIDs contain them. Technically we could strip the hyphens
out but the separator itself isn't important as long as it's not used in
ID.

Theoretically the proble existed before and still exists as it was
possible for two MaxScales to end up with the same random encryption key
and it's possible for two identical UUIDs to be generated on systems
with multiple parallel MaxScales. However this is is practically
impossible in real life.