MXS-4217: Accept larger keys in key_manager_file
The key manager should be able to handle keys of larger than what
MariaDB supports. As mxb::Cipher only supports certain keys, it should
handle the case where it is given one gracefully. Using exceptions works
in this case as all the code that uses it ends up catching the thrown
exception.