Use lower CPU count for test VMs
The backends don't need to be this big for testing and they can be
lowered down to two cores. The tests hard-code the `threads` value to 4
so anything above that is sort of wasted. The values could be lowered
even more to reduce the amount of waste capacity used by tests but this
is a good starting point.