MXS-4225: Fix maxlog log data parsing
The /v1/maxscale/logs/data endpoint incorrectly used raw line numbers as
the curson position but only returned the logical log lines (i.e. ones
with a timestamp and a log level) in the actual response. To fix this
and the creation of the cursors, the logical log lines should be
counted, not the raw ones.

This also fixes the REST API unit tests that were broken by MXS-4192
when it turned off syslog by default.