MXS-4206 Rename Max Rows to `Row limit`
MaxScale internally use the value of `max_rows` to set value for
`sql_select_limit`, so it works like a `LIMIT` clause and the users
have been used to with it. Therefore, naming it as `Max Rows`
would confuse them.

This also renames the components and its persisted state. i.e.
`query_max_rows` to `query_row_limit`.