Allow different `active_db` data for sessions
Having the same active db for all session tabs may be not what the
user wants despite the fact that MySql Workbench does it. It needs
to also detect if the user changes the active db, it needs to update
active db for all sessions of the worksheet.

Each session will have its `active_db` value. So when creating a new
session, the `active_db` will be empty.

This moves `active_db` from `schemaSidebar` module to `queryConn` module
as it's a part of the connection even though it's used to bold the
active db in the schemas sidebar.